Added: throw User Friendly Errors

+++ src/main/java/com/bridgelabz/employeepayrollapp/service/EmployeePayrollService.java
@@ -4,6 +4,7 @@ package com.bridgelabz.employeepayrollapp.service;
  */
 
 import com.bridgelabz.employeepayrollapp.DTO.EmployeeDTO;
+import com.bridgelabz.employeepayrollapp.Exeption.EmployeePayrollException;
 import com.bridgelabz.employeepayrollapp.model.Employee;
 import com.bridgelabz.employeepayrollapp.repository.EmployeePayrollRepository;
 
@@ -14,21 +15,13 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+
+//Created EmployeePayrollService class to serve api calls done by controller layer
 public class EmployeePayrollService implements IEmployeePayrollService {
+
+    //Autowired EmployeePayrollRepository interface to inject its dependency here
     @Autowired
     EmployeePayrollRepository repository;
-
-    public String getMessage(String name) {
-        return "Welcome " + name;
-    }
-
-    public String postMessage(EmployeeDTO employeeDTO) {
-        return "Hello " + employeeDTO.getFirstName() + "" + employeeDTO.getLastName() + "!";
-    }
-
-    public String putMessage(String name) {
-        return "How are you, " + name;
-    }
 
     public String getWelcome() {
         return "Welcome to Employee Payroll !!!";
@@ -45,19 +38,31 @@ public class EmployeePayrollService implements IEmployeePayrollService {
         return list;
     }
 
-    public Optional<Employee> getDataById(Integer id) {
-        Employee newEmployee = repository.getById(id);
-        return Optional.of(newEmployee);
+    public Employee getDataById(Integer id) {
+        Optional<Employee> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            return newEmployee.get();
+        } else throw new EmployeePayrollException("Employee id not found");
     }
 
     public Employee updateDataById(Integer id, EmployeeDTO employeeDTO) {
-        Employee newEmployee = new Employee(id, employeeDTO);
-        repository.save(newEmployee);
-        return newEmployee;
+        Optional<Employee> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            Employee employee = new Employee(id, employeeDTO);
+            repository.save(employee);
+            return employee;
+        } else {
+            throw new EmployeePayrollException("Employee Not found");
+        }
     }
 
     public String deleteDataById(Integer id) {
-        repository.deleteById(id);
+        Optional<Employee> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new EmployeePayrollException("Employee Details not found");
+        }
         return null;
     }
 }
